requirements.txt
sphinx_rtd_theme>=0.3.1
pygments>=2.4.1
nbsphinx
setuptools
scipy
numpy
astropy>=4.0
matplotlib>=3.1.0
iminuit>=2.0.0
pymc3
pystan
emcee
docutils<0.18
